Python/requirements.txt
fastapi
uvicorn
selenium
Pillow
opencv-python
numpy
transformers
torch
# tensorflow
# keras
# huggingface_hub
# numpy
# requests